+++ src/main/java/kancho/realestate/utils/excel/storeaprtment/ApartmentStoreMapper.java
@@ -1,4 +1,4 @@
-package kancho.realestate.utils.storeaprtment;
+package kancho.realestate.utils.excel.storeaprtment;
 
 import java.util.List;
 import java.util.Optional;
